pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Colin Budd <colin.budd@capita.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.1"
requests = "^2.24.0"
gunicorn = "^20.0.4"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pylint = "^2.5"
pytest = "^6.1"

[virtualenvs]
create = true
in-project = false
